in developing..

+++ draw.py
@@ -12,7 +12,7 @@ x = []
 y = []
 line = file.readline()
 while line:
-    [x1, x2, x3, x4, y1, y2, y3, y4] = list(map(lambda x: int(x),  line.split()))
+    [x1, x2, x3, x4, y1, y2, y3, y4] = list(map(lambda x: float(x),  line.split()))
     x.extend([x1,x2,x3,x4])
     y.extend([y1,y2,y3,y4])
     line = file.readline()
